refactor: further refine user profile state management
- Extract `useConnectUserProfile` custom hook to encapsulate profile state and validation.
- Move `Gender` field into `UserSettings.tsx` for a cohesive form.
- Centralize `UserProfileState` and `HrZoneData` types in `types/connect.ts`.
- Fix UX issue where transient state was cleared on validation error.
- Update `ConnectPage.tsx` to use the new hook, significantly reducing its size.
- Update VRT snapshots for the Connect page.
