+++ app/client/connect/ConnectView.tsx
@@ -19,16 +19,8 @@ import WorkoutControls from './WorkoutControls'
 import ResetSection from './components/ResetSection'
 import { useState, useEffect } from 'react'
 import logger from '@/utils/logger'
-import { Gender } from '../../../types/core'
 import { WorkoutStatus } from '../../../types/workout'
-import { UserProfileState, HrZoneData } from './types'
-import {
-  FormControl,
-  FormLabel,
-  RadioGroup,
-  FormControlLabel,
-  Radio,
-} from '@mui/material'
+import { UserProfileState, HrZoneData } from '@/types/connect'
 
 interface ConnectViewProps {
   duration: string
@@ -139,32 +131,7 @@ export default function ConnectView({
         </Typography>
 
         {!showUserDetails ? (
-          <Stack spacing={2} sx={{ mb: 3 }}>
-            <UserSettings profile={userProfile} />
-            <FormControl component="fieldset">
-              <FormLabel component="legend">Gender</FormLabel>
-              <RadioGroup
-                row
-                aria-label="gender"
-                name="gender"
-                value={data.gender}
-                onChange={(e) =>
-                  userProfile.handlers.setGender(e.target.value as Gender)
-                }
-              >
-                <FormControlLabel
-                  value="MALE"
-                  control={<Radio />}
-                  label="Male"
-                />
-                <FormControlLabel
-                  value="FEMALE"
-                  control={<Radio />}
-                  label="Female"
-                />
-              </RadioGroup>
-            </FormControl>
-          </Stack>
+          <UserSettings profile={userProfile} />
         ) : (
           <Box
             sx={{
